Update messages.json
Updated message.json regarding the latest 'Additional Actions' option. Note: like 'URL\Text' (Testo) and 'Grid\List' (Elenco), it would be nice to have 'left', 'right' and 'back' for mouse buttons as independent lines as well. This time it seems to me that it is necessary for a decent translation. Sorry. Why:

- position (left): 'a sinistra' \ button (left): 'sinistro'
- folder (back): 'precedende' \ button (back) 'indietro'

This time 

Thanks